Update controllers doc
Emphasis on exiting Steam completely rather than closing it

Since DS4Windows doesn't always provide the right .NET redist in the app, added an additional instruction to tell users to use the one on the Github directly if following the instructions still give the .NET error. This was an issue for multiple people and this always solved the issue.

Added a section for when the controller doesn't work.

Tell users to make sure controller is enabled, to try the controller in a game and that they followed the Steam instructions since we frequently see people who didn't read those instructions.

Also gave instructions for DS4Windows (that were both tested and also noticed lot of times in the support channels)

Last, added the exec default_controller.cfg command for World at War to fix controller not working as this has solved controller not working only on WAW for lot of people that did everything right