Update update doc
Removed the part that explained how to find the version in the changelog.
It used to be r2158-r2190 but it now only shows the new version and what it brings so this is outdated.

Removed BR html tag and use correct markdown instead

Reflect those changes in update.fr.md