loading-mods T6 instructions (compiler)
Removed notions of the compiler, of compiled files and requirements.

The T6 instructions are now the same than for IW5.
The only difference is the explanation to load scripts only in mp or only in zm and also how you can load both source and compiled scripts the same way (to not confuse people that would download older compiled scripts)

Added notes about not needing to restart the entire game to reload scripts
Added instructions to use map_restart

The iprintlnbold in the example was replaced with a simple message because the % and / were completely messing what was displayed on screen