Update moving-to-another-drive
Added a note to explain that this tutorial is for WAW specifically but that it works for any other folder

Rewrote the instructions because they were incorrect. The command would fail because the folder already exists. Now we tell the user to move the entire mods folder, tested on both W11 and W10

Made the Steam instructions clearer since some people have no clue that using exe installers installs mods "in Steam"

Added screenshots to show before and after

Explained that the two folders will stay synced, no need to re-run the command unless deleted